# contracts/agent-runtime.openapi.json
{
  "components": {
    "schemas": {
      "AlertResolvedEventPayload": {
        "additionalProperties": false,
        "properties": {
          "alertStatus": {
            "const": "RESOLVED",
            "title": "Alertstatus",
            "type": "string"
          },
          "endsAt": {
            "pattern": "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}\\.\\d{9}Z$",
            "title": "Endsat",
            "type": "string"
          },
          "incidentId": {
            "format": "uuid",
            "title": "Incidentid",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "runId": {
            "format": "uuid",
            "title": "Runid",
            "type": "string"
          },
          "runKind": {
            "$ref": "#/components/schemas/RunKind"
          },
          "schemaVersion": {
            "const": 5,
            "title": "Schemaversion",
            "type": "integer"
          }
        },
        "required": [
          "schemaVersion",
          "incidentId",
          "runId",
          "occurredAt",
          "runKind",
          "alertStatus",
          "endsAt"
        ],
        "title": "AlertResolvedEventPayload",
        "type": "object"
      },
      "AlertResolvedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AlertResolvedEventPayload"
          },
          "event": {
            "const": "alert.resolved",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "AlertResolvedStreamEvent",
        "type": "object"
      },
...
      "DiagnosisCompletedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DiagnosisCompletedEventPayload"
          },
          "event": {
            "const": "diagnosis.completed",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "DiagnosisCompletedStreamEvent",
        "type": "object"
      },
...
      "DiagnosisInsufficientStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DiagnosisInsufficientEventPayload"
          },
          "event": {
            "const": "diagnosis.insufficient",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "DiagnosisInsufficientStreamEvent",
        "type": "object"
      },
...
      "ErrorDetail": {
        "additionalProperties": false,
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "retryable": {
            "title": "Retryable",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "message",
          "retryable"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "EvidenceRecordedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/EvidenceRecordedEventPayload"
          },
          "event": {
            "const": "evidence.recorded",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "EvidenceRecordedStreamEvent",
        "type": "object"
      },
...
      "IncidentCreatedEventPayload": {
        "additionalProperties": false,
        "properties": {
          "attempt": {
            "minimum": 1.0,
            "title": "Attempt",
            "type": "integer"
          },
          "incidentId": {
            "format": "uuid",
            "title": "Incidentid",
            "type": "string"
          },
          "incidentStatus": {
            "const": "RECEIVED",
            "title": "Incidentstatus",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "runId": {
            "format": "uuid",
            "title": "Runid",
            "type": "string"
          },
          "runKind": {
            "const": "diagnosis",
            "title": "Runkind",
            "type": "string"
          },
          "runStatus": {
            "const": "QUEUED",
            "title": "Runstatus",
            "type": "string"
          },
          "schemaVersion": {
            "const": 5,
            "title": "Schemaversion",
            "type": "integer"
          }
        },
        "required": [
          "schemaVersion",
          "incidentId",
          "runId",
          "occurredAt",
          "runKind",
          "attempt",
          "incidentStatus",
          "runStatus"
        ],
        "title": "IncidentCreatedEventPayload",
        "type": "object"
      },
      "IncidentCreatedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/IncidentCreatedEventPayload"
          },
          "event": {
            "const": "incident.created",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "IncidentCreatedStreamEvent",
        "type": "object"
      },
...
      "OperatorSessionResponse": {
        "properties": {
          "csrfToken": {
            "title": "Csrftoken",
            "type": "string"
          },
          "expiresAt": {
            "title": "Expiresat",
            "type": "integer"
          },
          "operatorRef": {
            "title": "Operatorref",
            "type": "string"
          }
        },
        "required": [
          "operatorRef",
          "expiresAt",
          "csrfToken"
        ],
        "title": "OperatorSessionResponse",
        "type": "object"
      },
...
      "RepairDryRunPassedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RepairDryRunPassedEventPayload"
          },
          "event": {
            "const": "repair.dry_run_passed",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "RepairDryRunPassedStreamEvent",
        "type": "object"
      },
...
      "RepairPatchReadyStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RepairPatchReadyEventPayload"
          },
          "event": {
            "const": "repair.patch_ready",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "RepairPatchReadyStreamEvent",
        "type": "object"
      },
...
      "RepairWaitingApprovalStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RepairWaitingApprovalEventPayload"
          },
          "event": {
            "const": "repair.waiting_approval",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "RepairWaitingApprovalStreamEvent",
        "type": "object"
      },
...
      "RunEventHistoryResponse": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunEventStreamItem"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          },
          "schemaVersion": {
            "const": 5,
            "default": 5,
            "title": "Schemaversion",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "RunEventHistoryResponse",
        "type": "object"
      },
      "RunEventStreamItem": {
        "discriminator": {
          "mapping": {
            "alert.resolved": "#/components/schemas/AlertResolvedStreamEvent",
            "diagnosis.completed": "#/components/schemas/DiagnosisCompletedStreamEvent",
            "diagnosis.insufficient": "#/components/schemas/DiagnosisInsufficientStreamEvent",
            "evidence.recorded": "#/components/schemas/EvidenceRecordedStreamEvent",
            "incident.created": "#/components/schemas/IncidentCreatedStreamEvent",
            "repair.dry_run_passed": "#/components/schemas/RepairDryRunPassedStreamEvent",
            "repair.patch_ready": "#/components/schemas/RepairPatchReadyStreamEvent",
            "repair.waiting_approval": "#/components/schemas/RepairWaitingApprovalStreamEvent",
            "run.failed": "#/components/schemas/RunFailedStreamEvent",
            "run.queued": "#/components/schemas/RunQueuedStreamEvent",
            "run.started": "#/components/schemas/RunStartedStreamEvent",
            "tool.failed": "#/components/schemas/ToolFailedStreamEvent",
            "tool.started": "#/components/schemas/ToolStartedStreamEvent"
          },
          "propertyName": "event"
        },
        "oneOf": [
          {
            "$ref": "#/components/schemas/IncidentCreatedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/RunQueuedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/RunStartedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/ToolStartedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/EvidenceRecordedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/ToolFailedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/DiagnosisCompletedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/DiagnosisInsufficientStreamEvent"
          },
          {
            "$ref": "#/components/schemas/RunFailedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/RepairPatchReadyStreamEvent"
          },
          {
            "$ref": "#/components/schemas/RepairDryRunPassedStreamEvent"
          },
          {
            "$ref": "#/components/schemas/RepairWaitingApprovalStreamEvent"
          },
          {
            "$ref": "#/components/schemas/AlertResolvedStreamEvent"
          }
        ],
        "title": "RunEventStreamItem"
      },
...
      "RunFailedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunFailedEventPayload"
          },
          "event": {
            "const": "run.failed",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "RunFailedStreamEvent",
        "type": "object"
      },
...
      "RunQueuedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunQueuedEventPayload"
          },
          "event": {
            "const": "run.queued",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "RunQueuedStreamEvent",
        "type": "object"
      },
...
      "RunStartedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/RunStartedEventPayload"
          },
          "event": {
            "const": "run.started",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "RunStartedStreamEvent",
        "type": "object"
      },
...
      "ToolFailedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToolFailedEventPayload"
          },
          "event": {
            "const": "tool.failed",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "ToolFailedStreamEvent",
        "type": "object"
      },
...
      "ToolStartedStreamEvent": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ToolStartedEventPayload"
          },
          "event": {
            "const": "tool.started",
            "title": "Event",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "event",
          "data"
        ],
        "title": "ToolStartedStreamEvent",
        "type": "object"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "in": "cookie",
        "name": "__Host-k8s-incident-session",
        "type": "apiKey"
      },
      "AlertmanagerBearer": {
        "description": "Shared bearer credential mounted in Alertmanager and Runtime.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "K8s Incident Agent Runtime",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/incidents/{incident_id}/events": {
      "get": {
        "operationId": "get_incident_events_api_v1_incidents__incident_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Last-Event-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Last-Event-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/RunEventStreamItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Incident Events"
      }
    },
...
    "/api/v1/incidents/{incident_id}/runs/{run_id}/events": {
      "get": {
        "operationId": "list_run_events_api_v1_incidents__incident_id__runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "incident_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Incident Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunEventHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "List Run Events"
      }
    },
...
    "/api/v1/operator/login": {
      "post": {
        "operationId": "login_api_v1_operator_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "password": {
                    "format": "password",
                    "maxLength": 1024,
                    "minLength": 1,
                    "title": "Password",
                    "type": "string",
                    "writeOnly": true
                  }
                },
                "required": [
                  "password"
                ],
                "title": "OperatorLoginRequest",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Login"
      }
    },
    "/api/v1/operator/logout": {
      "post": {
        "operationId": "logout_api_v1_operator_logout_post",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Logout"
      }
    },
    "/api/v1/operator/session": {
      "get": {
        "operationId": "get_session_api_v1_operator_session_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Get Session"
      },
      "post": {
        "operationId": "renew_session_api_v1_operator_session_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatorSessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "summary": "Renew Session"
      }
    },
...
  }
}